Moved confirmation dialog to activity
